chore(sdam): create sdam subfolder, split topology up
